requirements.txt
Django==1.7.4
Pillow==2.7.0
redis==2.10.3
hiredis==0.1.6
django-redis-cache==0.13.0
argparse==1.3.0
djangorestframework==3.1.1
markdown==2.6.1
django-filter==0.9.2
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.1.1
psycopg2==2.6
static==0.4
wsgiref==0.1.2